cancellation request happens

+++ demo/src/main/java/com/example/demo/dto/response/FullAnalyticsResponse.java
@@ -1,11 +1,12 @@
 package com.example.demo.dto.response;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record FullAnalyticsResponse(
     BusinessTotalsDto businessTotals,
     List<MonthlyStatsDto> monthlyStats,
-    Double averageRating,
+    BigDecimal averageRating,
     List<ServiceDistributionDto> serviceDistribution,
     List<PeakHourDto> peakHour
 ) {}
